establish exception handling in controller. create package helpers with ErrorMessage. move IdService to namespace helpers.
