+++ backend/src/main/java/de/gregorstallmeister/backend/controller/GlobalExceptionHandler.java
@@ -1,0 +1,27 @@
+package de.gregorstallmeister.backend.controller;
+
+import de.gregorstallmeister.backend.helpers.ErrorMessage;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.time.Instant;
+import java.util.NoSuchElementException;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(NoSuchElementException.class)
+    public ErrorMessage handleNoSuchElementException(Exception ex) {
+        return new ErrorMessage("Ein Fehler ist aufgetreten: " + ex.getMessage(), Instant.now());
+    }
+
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(Exception.class)
+    public ErrorMessage handleException(Exception ex) {
+
+        return new ErrorMessage("Ein nicht bekannter Fehler ist aufgetreten: " + ex.getMessage(), Instant.now());
+    }
+}
