+++ backend/src/main/java/de/gregorstallmeister/backend/helpers/ErrorMessage.java
@@ -1,0 +1,9 @@
+package de.gregorstallmeister.backend.helpers;
+
+import java.time.Instant;
+
+public record ErrorMessage(
+        String message,
+        Instant instant
+) {
+}
