+++ backend/src/test/java/de/gregorstallmeister/backend/test/service/PictureServiceTest.java
@@ -1,5 +1,6 @@
 package de.gregorstallmeister.backend.test.service;
 
+import de.gregorstallmeister.backend.helpers.IdService;
 import de.gregorstallmeister.backend.model.*;
 import de.gregorstallmeister.backend.repository.PictureRepository;
 import de.gregorstallmeister.backend.service.PictureService;
